pyproject.toml
[project]
name = "fasthtml-instructor-labeling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.5.1",
    "pydantic>=2.9.0",
    "instructor>=1.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
